openvpn: remove legacy debian_openssl_vulnkeys patch
Remove the debian_openssl_vulnkeys patch as it's no longer included
in the Debian packaging for 2.2.3 and conflicts with changes in the
upstream OpenVPN 2.3 branch.  This was only included in Debian versions
at the time due to a Debian specific change to the OpenSSL package,
resulting in a predictable random number generator, and the subsequent
possibility of vulnerable keys.